requirements.txt
customtkinter==4.3.0
darkdetect==0.6.0
inflection==0.5.1
isort==5.10.1
Pillow==9.1.1
